add password by user

+++ app/Listeners/CheckAndDeletePreviousExam.php
@@ -3,6 +3,7 @@
 namespace App\Listeners;
 
 use App\Events\ExamCreated;
+use App\Models\Exam;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
 
@@ -22,10 +23,17 @@ class CheckAndDeletePreviousExam
     public function handle(ExamCreated $event): void
     {
 
-        $user = $event->exam->user;
-        $previousExam = $user->exams()->first();
-        if ($previousExam) {
+        $user_id = $event->exam->user_id;
+        // بررسی وجود امتحان قبلی
+        $previousExam = Exam::where('user_id', $user_id)->first();
+//dd($event->exam->id);
+        // اگر کاربر قبلاً امتحان ایجاد کرده باشد، امتحان جدید را حذف کنید
+        if ($previousExam && $previousExam->id !== $event->exam->id) {
             $previousExam->delete();
+        }
+        // اگر کاربر هنوز امتحانی ایجاد نکرده باشد، امتحان جدید را ذخیره کنید
+        if (!$previousExam) {
+            $event->exam->save();
         }
     }
 }
